src/test/java/de/cfranzen/adventofcode/day6/DataStreamBufferTest.java
package de.cfranzen.adventofcode.day6;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;

import static org.assertj.core.api.Assertions.assertThat;

class DataStreamBufferTest {

    @ParameterizedTest
    @CsvSource({
            "mjqjpqmgbljsphdztnvjfqwrcgsmlb,7",
            "bvwbjplbgvbhsrlpgdmjqwftvncz,5",
            "nppdvjthqldpwncqszvftbrmjlhg,6",
            "nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg,10",
            "zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw,11"
    })
    void testFindingStartOfPacketMarker(String data, int expectedMarker) {
        final int marker = new DataStreamBuffer(data).findFirstMarker(4);
        assertThat(marker).isEqualTo(expectedMarker);
    }

    @ParameterizedTest
    @CsvSource({
            "mjqjpqmgbljsphdztnvjfqwrcgsmlb,19",
            "bvwbjplbgvbhsrlpgdmjqwftvncz,23",
            "nppdvjthqldpwncqszvftbrmjlhg,23",
            "nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg,29",
            "zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw,26"
    })
    void testFindingStartOfMessageMarker(String data, int expectedMarker) {
        final int marker = new DataStreamBuffer(data).findFirstMarker(14);
        assertThat(marker).isEqualTo(expectedMarker);
    }
}